Add `#[clippy::print_hir]` attribute for debugging
